Move report bug in general category
