+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/composable/category/DebugCategory.kt
@@ -2,7 +2,6 @@ package com.flipperdevices.settings.impl.composable.category
 
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.flipperdevices.core.preference.pb.Settings
 import com.flipperdevices.core.ui.composable.LocalRouter
@@ -50,11 +49,5 @@ fun ColumnScope.DebugCategory(
         descriptionId = R.string.debug_shake2report_desc,
         state = settings.shakeToReport,
         onSwitchState = debugViewModel::onSwitchShakeToReport
-    )
-
-    val context = LocalContext.current
-    SimpleElement(
-        titleId = R.string.debug_shake2report_open,
-        onClick = { debugViewModel.onReportBug(context) }
     )
 }
