+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/viewmodels/SettingsViewModel.kt
@@ -1,11 +1,14 @@
 package com.flipperdevices.settings.impl.viewmodels
 
+import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.flipperdevices.core.di.ComponentHolder
+import com.flipperdevices.core.navigation.global.CiceroneGlobal
 import com.flipperdevices.core.preference.pb.Settings
 import com.flipperdevices.settings.impl.di.SettingsComponent
+import com.flipperdevices.shake2report.api.Shake2ReportApi
 import javax.inject.Inject
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
@@ -15,6 +18,12 @@ import kotlinx.coroutines.launch
 class SettingsViewModel : ViewModel() {
     @Inject
     lateinit var dataStoreSettings: DataStore<Settings>
+
+    @Inject
+    lateinit var cicerone: CiceroneGlobal
+
+    @Inject
+    lateinit var shakeToReportApi: Shake2ReportApi
 
     init {
         ComponentHolder.component<SettingsComponent>().inject(this)
@@ -47,6 +56,13 @@ class SettingsViewModel : ViewModel() {
                     .setEnabledExperimentalFunctions(value)
                     .build()
             }
+        }
+    }
+
+    fun onReportBug(context: Context) {
+        val screen = shakeToReportApi.reportBugScreen(context)
+        if (screen != null) {
+            cicerone.getRouter().navigateTo(screen)
         }
     }
 }
